requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==21.6b0
boto3==1.17.105
botocore==1.20.105
click==8.0.1
dataclasses==0.8
Django==3.1.7
django-ses==2.0.0
djangorestframework==3.12.2
future==0.18.2
importlib-metadata==4.6.0
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.8.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.7.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.4.1
